add readings

+++ src/controllers/fileAccess/index.ts
@@ -1,15 +1,15 @@
 import { Request, Response } from 'express';
 import { FileAccessState } from './fileAccess.model';
 import _ from 'lodash';
-import { fileAccessValidation } from './fileAcess.validation';
+import { fileAccessValidation, removeFileValidation } from './fileAcess.validation';
 import { FileAccess } from '../../models/fileAccess';
 import { sendResponse } from '../../helpers/sendResponse';
 import { User } from '../../models/user';
 import { File } from '../../models/file';
 
-const dataPicker = (data: any): FileAccessState => {
+const dataPicker = (data: any, keys?: string[]): FileAccessState => {
 
-    return _.pick(data, ['user_id', 'file_id', 'access_type'])
+    return _.pick(data, Array.isArray(keys) && keys?.length > 0 ? keys : ['user_id', 'file_id', 'access_type'])
 
 }
 
@@ -36,10 +36,14 @@ export const giveFileAccess = async (req: Request, res: Response): Promise<void>
 
 export const removeFileAccess = async (req: Request, res: Response): Promise<void> => {
 
-    let data: FileAccessState = dataPicker(req.body);
+    const data: FileAccessState = dataPicker(req.body, ['user_id', 'file_id']);
+    const userId = req.headers.userId;
 
-    let { error } = await fileAccessValidation(data);
+    const { error } = await removeFileValidation(data);
     if (error) throw error;
+
+    const file = await File.findOne({ _id: data.file_id, user_id: userId });
+    if (!file) { return sendResponse(res, 400, 'File not found or you are not the owner'); }
 
     await FileAccess.deleteOne({ user_id: data.user_id, file_id: data.file_id });
     sendResponse(res, 200, 'Access removed successfully');
